doc fix, float to int

+++ 0x06-python-classes/3-square.py
@@ -29,6 +29,6 @@ class Square:
         Calculate the area of the square.
 
         Returns:
-            float: The area of the square.
+            int: The area of the square.
         """
         return self.__size ** 2
